[bench] oneRound not to redraw GLbsed benchmarks

+++ src/org/zeroxlab/benchmark/CaseGLCube.java
@@ -41,7 +41,7 @@ public class CaseGLCube extends Case{
 	int length = mResult.length;
 
 	for (int i = 0; i < length; i++) {
-	    long fps = mCaseRound / (mResult[i] / 1000); // milliseconds to seconds
+	    float fps = mCaseRound / (mResult[i] / 1000f); // milliseconds to seconds
 	    result += "Round " + i + ": fps = " + fps + "\n";
 	    total  += fps;
 	}
